First try to refactor
Something is not yes in the PermissionsFragment.kt and CameraFragment.kt.
