+++ Android/PolishBanknotesApp/README.md
@@ -1,1 +1,4 @@
 # Polish banknotes Android app
+The project is based on https://github.com/tensorflow/examples/tree/master/lite/examples/image_classification/android
+on the license shown in variuos files (http://www.apache.org/licenses/LICENSE-2.0). 
+Every change has been made or approved by me (https://github.com/jankejc).
