+++ Android/PolishBanknotesApp/app/build.gradle
@@ -13,6 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+/*
+* The project is based on https://github.com/tensorflow/examples/tree/master/lite/examples/image_classification/android
+* on above license. Every change has been made or approved by me (https://github.com/jankejc).
+*/
 
 plugins {
     id 'com.android.application'
@@ -23,10 +27,11 @@ plugins {
 }
 
 android {
+    namespace 'pg.eti.project.polishbanknotes'
     compileSdk 32
 
     defaultConfig {
-        applicationId "org.tensorflow.lite.examples.imageclassification"
+        applicationId "pg.eti.project.polishbanknotes"
         minSdk 23
         targetSdk 32
         versionCode 1
