--- Android/PolishBanknotesApp/app/download_models.gradle
@@ -1,47 +1,0 @@
-/*
- * Copyright 2022 The TensorFlow Authors. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *             http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-task downloadModelFile(type: Download) {
-    src 'https://storage.googleapis.com/download.tensorflow.org/models/tflite/task_library/image_classification/android/mobilenet_v1_1.0_224_quantized_1_metadata_1.tflite'
-    dest project.ext.ASSET_DIR + '/mobilenetv1.tflite'
-    overwrite false
-}
-
-task downloadModelFile0(type: Download) {
-    src 'https://storage.googleapis.com/download.tensorflow.org/models/tflite/task_library/image_classification/android/efficientnet_lite0_int8_2.tflite'
-    dest project.ext.ASSET_DIR + '/efficientnet-lite0.tflite'
-    overwrite false
-}
-
-task downloadModelFile1(type: Download) {
-    src 'https://storage.googleapis.com/download.tensorflow.org/models/tflite/task_library/image_classification/android/efficientnet_lite1_int8_2.tflite'
-    dest project.ext.ASSET_DIR + '/efficientnet-lite1.tflite'
-    overwrite false
-}
-
-task downloadModelFile2(type: Download) {
-    src 'https://storage.googleapis.com/download.tensorflow.org/models/tflite/task_library/image_classification/android/efficientnet_lite2_int8_2.tflite'
-    dest project.ext.ASSET_DIR + '/efficientnet-lite2.tflite'
-    overwrite false
-}
-
-task copyTestModel(type: Copy, dependsOn: downloadModelFile) {
-    from project.ext.ASSET_DIR + '/mobilenetv1.tflite'
-    into project.ext.TEST_ASSETS_DIR
-}
-
-preBuild.dependsOn downloadModelFile, downloadModelFile0, downloadModelFile1, downloadModelFile2,
-        copyTestModel
