+++ Android/PolishBanknotesApp/app/src/androidTest/java/pg/eti/project/polishbanknotes/ImageClassificationTest.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.tensorflow.lite.examples.imageclassification
+package pg.eti.project.polishbanknotes.imageclassification
 
 import android.content.res.AssetManager
 import android.graphics.Bitmap
