+++ Android/PolishBanknotesApp/app/src/main/res/values/colors.xml
@@ -13,6 +13,9 @@
     ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
     ~ See the License for the specific language governing permissions and
     ~ limitations under the License.
+    ~ NOTE:
+    ~ The project is based on https://github.com/tensorflow/examples/tree/master/lite/examples/image_classification/android
+    ~ on above license. Every change has been made or approved by me (https://github.com/jankejc).
 -->
 
 <resources>
