+++ Android/PolishBanknotesApp/app/src/main/res/values/strings.xml
@@ -12,9 +12,12 @@
         ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
         ~ See the License for the specific language governing permissions and
         ~ limitations under the License.
+        ~ NOTE:
+        ~ The project is based on https://github.com/tensorflow/examples/tree/master/lite/examples/image_classification/android
+        ~ on above license. Every change has been made or approved by me (https://github.com/jankejc).
     -->
 <resources>
-    <string name="app_name">TFLite Image Classification</string>
+    <string name="app_name">Polskie Banknoty</string>
 
     <string name="alt_bottom_sheet_chevron">Bottom sheet expandable indicator</string>
