+++ Android/PolishBanknotesApp/settings.gradle
@@ -12,5 +12,5 @@ dependencyResolutionManagement {
         mavenCentral()
     }
 }
-rootProject.name = "TFLite Image Classification"
+rootProject.name = "Polish Banknotes"
 include ':app'
